change this to int64 cause it fixes the intermission thing on high res
idk why that fixes it tbh but it does

+++ src/y_inter.c
@@ -437,7 +437,7 @@ void Y_IntermissionDrawer(void)
 
 	if (sorttic != -1 && intertic >= sorttic && !demo.playback)
 	{
-		INT32 count = (intertic - sorttic);
+		INT64 count = (intertic - sorttic);
 
 		if (count < 8)
 			x -= ((((count<<FRACBITS) + R_GetHudUncap()) * vid.width)>>FRACBITS) / (8 * vid.dupx);
